requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
django-cors-headers==3.11.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
Django==4.0.1
djangorestframework-simplejwt==4.8.0
djangorestframework==3.13.1
djoser==2.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.1.1
pgsql==1.2
pip==20.0.2
pkg-resources==0.0.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests-oauthlib==1.3.0
requests==2.27.1
setuptools==44.0.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
wheel==0.34.2
